Creating SessionHelper module to dry out spec tests
