+++ spec/features/sign_in_spec.rb
@@ -1,13 +1,7 @@
 feature 'Signing in' do
 
   scenario 'exisitng user can give their username and password' do
-    user = User.create(user_params)
-    visit '/'
-    within 'form#sign_in' do
-      fill_in 'username', with: user.username
-      fill_in 'password', with: user.password
-    end
-    click_button 'Sign in'
+    sign_in
     expect(page).to have_content 'Hi test, here are your Peeps'
   end
 
@@ -32,13 +26,6 @@ feature 'Signing in' do
     end
     click_button 'Sign in'
     expect(page).to have_content 'Hello world!'
-  end
-
-  def user_params
-    { email: 'example@example.com',
-      password: 'secret123',
-      name: 'test',
-      username: 'test_name' }
   end
 
 end
