+++ spec/features/sign_out_spec.rb
@@ -1,22 +1,9 @@
 feature 'Signing out' do
 
   scenario 'after logging in a user can sign out' do
-    user = User.create(user_params)
-    visit '/'
-    within 'form#sign_in' do
-      fill_in 'username', with: user.username
-      fill_in 'password', with: user.password
-    end
-    click_button 'Sign in'
+    sign_in
     click_button 'Sign out'
     expect(page).to have_content 'Goodbye!'
-  end
-
-  def user_params
-    { email: 'example@example.com',
-      password: 'secret123',
-      name: 'test',
-      username: 'test_name' }
   end
 
 end
